feat: added curriculum entities
