--- src/main/java/com/university/librarymanagementsystem/entity/Department.java
@@ -1,24 +1,0 @@
-package com.university.librarymanagementsystem.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "department")
-@Data
-public class Department {
-
-    @Id
-    private int id;
-    private String name;
-}
