+++ src/main/java/com/university/librarymanagementsystem/entity/curriculum/Program.java
@@ -1,0 +1,40 @@
+package com.university.librarymanagementsystem.entity.curriculum;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "program")
+@Data
+public class Program {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @ManyToOne
+    @JoinColumn(name = "department_id", nullable = false)
+    private Department department;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+    @Column(name = "subjects", nullable = false)
+    private int numberOfSubjects;
+    @Column(name = "status", nullable = false)
+    private int status;
+}
