+++ src/main/java/com/university/librarymanagementsystem/entity/curriculum/Subject.java
@@ -1,12 +1,12 @@
-package com.university.librarymanagementsystem.entity;
-
-import org.hibernate.annotations.Collate;
+package com.university.librarymanagementsystem.entity.curriculum;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -21,13 +21,25 @@ import lombok.Setter;
 @Entity
 @Table(name = "subjects")
 @Data
-public class Subjects {
+public class Subject {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
+
+    @ManyToOne
+    @JoinColumn(name = "program_id", nullable = false)
+    private Program program;
+
+    @Column(name = "department_id")
+    private int department_id;
 
     @Column(name = "subject_name")
-    private String subjectName;
+    private String name;
 
+    @Column(name = "year", nullable = false)
+    private int year;
+
+    @Column(name = "status", nullable = false)
+    private int status;
 }
